+++ bindings/deno/data_manifest.ts
@@ -61,4 +61,3 @@ export const MORPHEUS_DATASETS: Readonly<
 };
 
 export const MORPHEUS_DATA_SCHEMA_VERSION = 1;
-export const MORPHEUS_PACKAGE_VERSION = "0.3.2";